Added literal and plus to parser/scanner/ast
Glut stuf is now working and compiling (not sure what fixed it, but I no
longer get the CR/LF error). Please see if you can get it to work on
yours. Need to add a file that is equivilant to bytecode.ml of microC.
Maybe cflat.ml or something. Assuming we don't need an interperter (for
diagnostics), we need to restructure compile.ml to read from flatc.ml.